requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-extensions
psycopg2
django-cors-headers